Fixed exception on non existing config file on MacOS test tear down

+++ DataSpace.Tests/Common/Settings/Connection/AccountSettingsTest.cs
@@ -49,7 +49,7 @@ namespace Tests.Common.Settings.Connection {
 
         [TearDown]
         public void CleanUp() {
-            if (configuration.HasFile) {
+            if (configuration.HasFile && File.Exists(configuration.FilePath)) {
                 File.Delete(configuration.FilePath);
             }
         }
